First Commit Currently Working on PC adding [DONE], NPC adding[DONE].
Then Selection Of PCs and NPCs using tabulate to tidy it up nice
Started this part with just the displays[DONE], will add in a selector[DONE], but will try to make it nuetral[DONE] on which list it's pulling from
Everything but the checks is working.  Tabulate is awesome.
CURRENTLY: Trying to get checks running, and data collected

IDEAS FOR FURTHER VERSIONS:
1)Make me able to use tool to run stealth checks against PCs too (Result in a single (Character)class instead of two(PC, and NPC) most likely, maybe not though.. we'll see.)
2A)Either keep a file of the Characters that are present (load at beginning, save at end) Which obviosly means amke a way of deleteing entries
2B) or set it up in a DB with entries and such using Peewee and SQLITE